LandscapeLeadsFL/requirements.txt
# Core dependencies
pandas>=2.0.0
beautifulsoup4>=4.12.0
patchright>=0.1.0  # Patched Playwright for stealth browsing

# Optional: Google search (may have rate limits)
googlesearch-python>=1.2.0

# HTTP client (fallback)
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
lxml>=5.0.0
